refactor: clean observations code
